Update index.md
Corrected the incorrectly used word 'thought' to 'through'. 
Corrected grammar of 'how to write an extensions'. 
Clarified the meaning of 'ensuring that you really need to start coding'.
Added full stops to 'aka' abbreviation to make it clearer it is an abbreviation.
Changed 'developer environment' to more conventional phrase 'development environment'.